runner/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.27.2
beautifulsoup4==4.13.4
pydantic==2.11.7
playwright==1.55.0
qdrant-client==1.12.1
openai==1.66.3
langchain-text-splitters==0.3.8
python-dotenv==1.0.1
langchain-community==0.3.19
langchain==0.3.20
langchain-openai==0.3.7
langchain-qdrant==0.2.0
langgraph==0.2.67
ragas==0.2.14
rapidfuzz==3.9.7
